requirements.txt
fastapi==0.70.0
Jinja2==3.0.2
pandas==1.3.4
pydantic==1.8.2
tortoise-orm==0.17.8
uvicorn==0.15.0
